refactor: remove console logs and update default folded state in table components and add fold on table node
